// src/types/product.type.ts
export interface Product {
  _id?: string
  images?: string[]
  price?: number
  rating?: number
  price_before_discount?: number
  quantity?: number
  sold?: number
  view?: number
  name?: string
  description?: string
  category?: {
    _id?: string
    name?: string
  }
  image?: string
  createdAt?: string
  updatedAt?: string
}

export interface ProductList {
  message?: string
  data?: {
    products?: Product[]
    pagination?: {
      page?: number
      limit?: number
      page_size?: number
    }
  }
}

export interface ProductConfig {
  page?: number
  limit?: number
  sort_by?: 'createdAt' | 'view' | 'sold' | 'price'
  order?: 'desc' | 'asc'
  category?: string
  exclude?: string
  rating_filter?: number
  price_max?: number
  price_min?: number
  name?: string
}